Rename softkbdgen to kbdgen
As the scope of the project evolved, the name softkbdgen seems less and
less appropriate. Originally, the project only output keyboards for
Android and iOS.

It now generates outputs for X11, OS X and Windows, and also supports
generic SVG output for development. It also supports more tooling to
convert CLDR XML files of keyboards to the kbdgen YAML format.

We're evolving, and v1.0 comes soon! :D