Collection of changes from VM
A collection of changes from testing this on a Ubuntu VM. These changes
are the squashed version of testing MIM exporting live on an up-to-date
Ubuntu 19.04. The generated files work for the most part -- I couldn't
manage to get super and hyper layers to correctly work, and I've only
tested it with German and US keyboards.

Please note that for the grouping to work, the `mim.yaml` target file
should be used. Further explanations are `src/m17n_mim/Readme.md`.